+++ source/expect/ToBeAssignableTo.ts
@@ -1,6 +1,7 @@
-import { RelationMatcherBase } from "./RelationMatcherBaseNew.js";
+import { RelationMatcherBase } from "./RelationMatcherBase.js";
 
 export class ToBeAssignableTo extends RelationMatcherBase {
   relation = this.typeChecker.relation.assignable;
   relationExplanationText = "assignable to";
+  relationExplanationVerb = "is";
 }
