+++ source/expect/ToEqual.ts
@@ -3,4 +3,5 @@ import { RelationMatcherBase } from "./RelationMatcherBase.js";
 export class ToEqual extends RelationMatcherBase {
   relation = this.typeChecker.relation.identity;
   relationExplanationText = "identical to";
+  relationExplanationVerb = "is";
 }
