+++ source/expect/ToMatch.ts
@@ -2,5 +2,6 @@ import { RelationMatcherBase } from "./RelationMatcherBase.js";
 
 export class ToMatch extends RelationMatcherBase {
   relation = this.typeChecker.relation.subtype;
-  relationExplanationText = "a subtype of";
+  relationExplanationText = "match";
+  relationExplanationVerb = "does";
 }
